// src/app/pages/transfer/setting-transfer/list-setting/shared/list-setting.enum.ts
export enum ListSettingEnum {
   vehicleFeatureTitle = 'Vehicle',
   haveChildrenTitle = 'Children',
   allowPetsTitle = 'Pets',
   doorToDoorTitle = 'Door to Door',
   stopsOrPickupsTitle = 'Stops or Pickups',
   transferDriverTitle = 'Driver',
   actionTitle = 'Action',

   vehicleDescription = 'vehicleDescription',
   haveChildren = 'haveChildren',
   allowPets= 'allowPets',
   doorToDoor = 'doorToDoor',
   stopsOrPickups = 'stopsOrPickups',
   driverName = 'driverName',
   action = 'action',  
  
   valueOne = 1,
   valueTwo = 2,
}